more fluid terrain based on interpolation between samples

+++ MiniMinecraft/Shaders/terrain_header.metal
@@ -11,13 +11,13 @@ using namespace metal;
 
 #define M_PI 3.1415926535f
 
-uint8_t inSinWeightedTerrain(thread float3 pos);
+float inSinWeightedTerrain(thread float3 pos);
 
 uint8_t inCheckeredTerrain(thread float3 pos);
 
-uint8_t inSphereTerrain(thread float3 pos);
+float inSphereTerrain(thread float3 pos);
 
-uint8_t inPerlinTerrain(thread float3 pos);
+float inPerlinTerrain(thread float3 pos);
 
 uint8_t inFrameTerrain(thread float3 pos);
